Fix subtle bug with detection of current language
The previous code would return the device's language setting, however it is possible that the app does not have a localization for that setting.

The new code returns the language actually being used within the app, which may be a left-to-right language (e.g. English) even though the system setting is to a right-to-left language (e.g. Hebrew) because that system language isn't supported in the app. This change prevents the right-to-left flip from occurring in this case.