Add warnings to constraint removal convenience methods
Per a direct conversation with Apple, using these methods to "nuke" all the constraints on a view or view hierarchy and then installing a fresh set of constraints will result in disastrous performance when the Auto Layout engine/solver attempts to recalculate the view frames.

Instead, you'll want to make sure views are only reused with the same constraints (or with very small differences or changes to the constant value of the constraints).