Deprecate some API methods
- Rename autoCenterInSuperviewAlongAxis: to
autoAlignAxisToSuperviewAxis: to reduce confusion about whether it
works vertically or horizontally. The old method (with old name) lives
on temporarily for compatibility, but is deprecated.
- Deprecate the “absolute positioning” methods, which encourage bad
habits when thinking about constraints. There is always another
(usually better) way to accomplish the same thing, so in an effort to
simplify the API these are being removed.