Automatically set translatesAutoresizingMaskIntoConstraints to NO when constraining a view
When calling the API methods on a view that has
translatesAutoresizingMaskIntoConstraints set to YES (the default),
this will automatically be changed to NO so that auto layout can be
used.

Note that only the view receiving the message has this property
changed; if there is another view that this view is constrained to, its
translatesAutoresizingMaskIntoConstraints property will not be
affected. (This is done to preserve the ability to constrain views to
other views not using auto layout.) Likewise, the view that the
constraint is added to is not affected.